An Analysis of the 2023 Presidential Election
Here is an analysis of the 2023 Presidential Elections for wards and polling units in Lagos State. The analysis focuses on the voting behaviors and patterns in each polling units which in turn was used to determine the outliers and bias in the individual polling units for the major political parties.